remove actor blockstore package
