+++ Cargo.toml
@@ -29,7 +29,6 @@ fil_actor_power = { workspace = true, features = ["fil-actor"] }
 fil_actor_reward = { workspace = true, features = ["fil-actor"] }
 fil_actor_system = { workspace = true, features = ["fil-actor"] }
 fil_actor_verifreg = { workspace = true, features = ["fil-actor"] }
-fil_actor_blockstore = {workspace = true}
 
 [build-dependencies]
 fil_actor_bundler = "6.1.0"
@@ -58,8 +57,7 @@ members = [
      "runtime",
      "test_vm",
      "vm_api",
-     "integration_tests",
-     "actor_blockstore"
+     "integration_tests"
 ]
 
 [workspace.dependencies]
@@ -148,7 +146,6 @@ fil_builtin_actors_state = { version = "12.0.0", path = "state"}
 fil_actors_integration_tests = { version = "1.0.0", path = "integration_tests" }
 vm_api = { version = "1.0.0", path = "vm_api" }
 test_vm = { version = "12.0.0", path = "test_vm" }
-fil_actor_blockstore = { version = "0.1.0", path = "actor_blockstore"}
 
 [patch.crates-io]
 #fvm_shared = { git = "https://github.com/filecoin-project/ref-fvm", branch = "master" }
