--- actor_blockstore/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "fil_actor_blockstore"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-fvm_ipld_blockstore = { workspace = true }
-anyhow = { workspace = true }
-cid = { workspace = true }
-multihash = {workspace = true }
-
-[lib]
-## lib is necessary for integration tests
-## cdylib is necessary for Wasm build
-crate-type = ["cdylib", "lib"]
