--- actor_blockstore/src/lib.rs
@@ -1,4 +1,0 @@
-pub mod tracking;
-
-mod memory;
-pub use memory::MemoryBlockstore;
