+++ actors/evm/Cargo.toml
@@ -32,7 +32,6 @@ fil_actors_evm_shared = { workspace = true }
 hex = { workspace = true }
 hex-literal = { workspace = true }
 substrate-bn = { workspace = true }
-fil_actor_blockstore = {workspace = true}
 
 [dev-dependencies]
 hex = { workspace = true, features = ["serde"] }
