+++ actors/market/Cargo.toml
@@ -22,7 +22,6 @@ frc42_dispatch = { workspace = true }
 frc46_token = { workspace = true }
 fvm_ipld_bitfield = { workspace = true }
 fvm_ipld_blockstore = { workspace = true }
-fil_actor_blockstore = { workspace = true }
 fvm_ipld_encoding = { workspace = true }
 fvm_ipld_hamt = { workspace = true }
 fvm_shared = { workspace = true }
