+++ actors/miner/Cargo.toml
@@ -33,7 +33,6 @@ lazy_static = { workspace = true }
 log = { workspace = true }
 byteorder = { workspace = true }
 itertools = { workspace = true }
-fil_actor_blockstore = {workspace = true}
 
 [dev-dependencies]
 fil_actors_runtime = { workspace = true, features = ["test_utils", "sector-default"] }
