+++ actors/miner/tests/miner_actor_test_bitfield_queue.rs
@@ -1,5 +1,5 @@
-use fil_actor_blockstore::MemoryBlockstore;
 use fil_actor_miner::BitFieldQueue;
+use fil_actors_runtime::test_blockstores::MemoryBlockstore;
 use fil_actors_runtime::test_utils::MockRuntime;
 use fvm_ipld_amt::Amt;
 use fvm_ipld_bitfield::BitField;
