+++ actors/miner/tests/sectors.rs
@@ -1,6 +1,6 @@
-use fil_actor_blockstore::MemoryBlockstore;
 use fil_actor_miner::SectorOnChainInfo;
 use fil_actor_miner::Sectors;
+use fil_actors_runtime::test_blockstores::MemoryBlockstore;
 use fil_actors_runtime::test_utils::*;
 use fvm_ipld_bitfield::BitField;
 use fvm_shared::sector::RegisteredSealProof;
