+++ actors/miner/tests/util.rs
@@ -7,7 +7,7 @@ use std::ops::Neg;
 
 use cid::multihash::MultihashDigest;
 use cid::Cid;
-use fil_actor_blockstore::MemoryBlockstore;
+use fil_actors_runtime::test_blockstores::MemoryBlockstore;
 use fvm_ipld_amt::Amt;
 use fvm_ipld_bitfield::iter::Ranges;
 use fvm_ipld_bitfield::{BitField, UnvalidatedBitField, Validate};
