+++ runtime/src/test_blockstores.rs
@@ -1,13 +1,44 @@
 // Copyright 2021-2023 Protocol Labs
 // SPDX-License-Identifier: Apache-2.0, MIT
-//#![cfg(feature = "tracking")]
-
 use std::cell::RefCell;
+use std::collections::HashMap;
 
 use anyhow::Result;
-use cid::multihash::{self, Code};
 use cid::Cid;
+
 use fvm_ipld_blockstore::{Block, Blockstore};
+use multihash::Code;
+
+#[derive(Debug, Default, Clone)]
+pub struct MemoryBlockstore {
+    blocks: RefCell<HashMap<Cid, Vec<u8>>>,
+}
+
+impl MemoryBlockstore {
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    /// Copy all blocks from this blockstore into the target blockstore.
+    pub fn copy_to(&self, other: &impl Blockstore) -> Result<()> {
+        other.put_many_keyed(self.blocks.borrow().iter().map(|(&k, v)| (k, v)))
+    }
+}
+
+impl Blockstore for MemoryBlockstore {
+    fn has(&self, k: &Cid) -> Result<bool> {
+        Ok(self.blocks.borrow().contains_key(k))
+    }
+
+    fn get(&self, k: &Cid) -> Result<Option<Vec<u8>>> {
+        Ok(self.blocks.borrow().get(k).cloned())
+    }
+
+    fn put_keyed(&self, k: &Cid, block: &[u8]) -> Result<()> {
+        self.blocks.borrow_mut().insert(*k, block.into());
+        Ok(())
+    }
+}
 
 /// Stats for a [TrackingBlockstore] this indicates the amount of read and written data
 /// to the wrapped store.
@@ -107,12 +138,12 @@ where
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::MemoryBlockstore;
+    use crate::test_blockstores;
     use fvm_ipld_blockstore::Block;
 
     #[test]
     fn basic_tracking_store() {
-        let mem = MemoryBlockstore::default();
+        let mem = test_blockstores::MemoryBlockstore::default();
         let tr_store = TrackingBlockstore::new(&mem);
         assert_eq!(*tr_store.stats.borrow(), BSStats::default());
