+++ runtime/tests/mapmap_test.rs
@@ -1,5 +1,5 @@
 use cid::Cid;
-use fil_actor_blockstore::MemoryBlockstore;
+use fil_actors_runtime::test_blockstores::MemoryBlockstore;
 use fil_actors_runtime::{make_map_with_root_and_bitwidth, Keyer, Map, MapMap};
 use fvm_shared::HAMT_BIT_WIDTH;
