+++ runtime/tests/set_test.rs
@@ -5,7 +5,7 @@ use fil_actors_runtime::Set;
 
 #[test]
 fn put() {
-    let store = fil_actor_blockstore::MemoryBlockstore::new();
+    let store = fil_actors_runtime::test_blockstores::MemoryBlockstore::new();
     let mut set = Set::new(&store);
 
     let key = "test".as_bytes();
@@ -17,7 +17,7 @@ fn put() {
 
 #[test]
 fn collect_keys() {
-    let store = fil_actor_blockstore::MemoryBlockstore::new();
+    let store = fil_actors_runtime::test_blockstores::MemoryBlockstore::new();
     let mut set = Set::new(&store);
 
     set.put("0".into()).unwrap();
@@ -33,7 +33,7 @@ fn collect_keys() {
 
 #[test]
 fn delete() {
-    let store = fil_actor_blockstore::MemoryBlockstore::new();
+    let store = fil_actors_runtime::test_blockstores::MemoryBlockstore::new();
     let mut set = Set::new(&store);
 
     let key = "0".as_bytes();
