+++ test_vm/tests/suite/batch_onboarding_deals_test.rs
@@ -1,5 +1,5 @@
-use fil_actor_blockstore::MemoryBlockstore;
 use fil_actors_integration_tests::tests::batch_onboarding_deals_test;
+use fil_actors_runtime::test_blockstores::MemoryBlockstore;
 use test_vm::TestVM;
 
 #[test]
