+++ test_vm/tests/suite/init_test.rs
@@ -1,5 +1,5 @@
-use fil_actor_blockstore::MemoryBlockstore;
 use fil_actors_integration_tests::tests::placeholder_deploy_test;
+use fil_actors_runtime::test_blockstores::MemoryBlockstore;
 use test_vm::TestVM;
 
 #[test]
