+++ test_vm/tests/suite/replica_update_test.rs
@@ -1,6 +1,6 @@
-use fil_actor_blockstore::MemoryBlockstore;
 use fil_actors_integration_tests::tests::create_miner_and_upgrade_sector;
 use fil_actors_runtime::runtime::Policy;
+use fil_actors_runtime::test_blockstores::MemoryBlockstore;
 use test_case::test_case;
 use test_vm::TestVM;
