+++ test_vm/tests/suite/verified_claim_test.rs
@@ -1,7 +1,7 @@
-use fil_actor_blockstore::MemoryBlockstore;
 use fil_actors_integration_tests::tests::{
     deal_passes_claim_fails_test, expired_allocations_test, verified_claim_scenario_test,
 };
+use fil_actors_runtime::test_blockstores::MemoryBlockstore;
 use test_vm::TestVM;
 
 // Tests a scenario involving a verified deal from the built-in market, with associated
